+++ tests/DependentTypes.Tests/RunTests.fs
@@ -7,7 +7,11 @@ module RunTests =
     [<EntryPoint>]
     let main args =
 
-        Tests.runTestsWithArgs defaultConfig args DependentType.dependentTypes |> ignore
-        
-
-        0
+        [
+            Tests.runTestsWithArgs defaultConfig args DependentType.optionDependentType
+            Tests.runTestsWithArgs defaultConfig args DependentType.singleDependentType
+            Tests.runTestsWithArgs defaultConfig args DependentType.sumDependentType
+            Tests.runTestsWithArgs defaultConfig args DependentType.convertsToDependentType
+            Tests.runTestsWithArgs defaultConfig args DependentType.dependentPairs
+        ] 
+        |> List.sum