+++ tests/DomainLib.Tests/RunTests.fs
@@ -7,10 +7,11 @@ module RunTests =
     [<EntryPoint>]
     let main args =
 
-        Tests.runTestsWithArgs defaultConfig args DomainTypes.trimNonEmptyString |> ignore
-        Tests.runTestsWithArgs defaultConfig args DomainTypes.digits |> ignore
-        Tests.runTestsWithArgs defaultConfig args DomainTypes.digits2 |> ignore
-        Tests.runTestsWithArgs defaultConfig args DomainTypes.genericSet |> ignore
-
-        0
+        [
+            Tests.runTestsWithArgs defaultConfig args DomainTypes.trimNonEmptyString
+            Tests.runTestsWithArgs defaultConfig args DomainTypes.digits
+            Tests.runTestsWithArgs defaultConfig args DomainTypes.digits2
+            Tests.runTestsWithArgs defaultConfig args DomainTypes.genericSet
+        ]
+        |> List.sum